LayoutTests/imported/w3c/web-platform-tests/css/css-anchor-position/position-try-grid-001-expected.txt

1
2
3
4
5
7
8
9
10
11
12
13
14
15
16

FAIL .target 1 assert_equals:
<div class="target" data-offset-x="135" data-offset-y="70" data-expected-height="100"></div>
offsetLeft expected 135 but got 200
